requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Talisman==1.1.0
Werkzeug==2.3.7
segno==1.6.6
python-dotenv==1.2.1
gunicorn==21.2.0
email-validator==2.0.0
WTForms==3.0.1
whitenoise==6.6.0
markdown==3.10.1
psycopg2-binary==2.9.9
razorpay==1.4.1
requests==2.31.0
Flask-Migrate==4.0.5